remove codes
